# main-service/src/main/resources/api-spec.yaml
openapi: 3.0.1
info:
  title: Main service API
  version: '1.0'
servers:
  - description: Generated server url
    url: http://localhost:8080
...
paths:
...
  /admin/events:
    get:
      description: |-
        Эндпоинт возвращает полную информацию обо всех событиях подходящих под переданные условия

        В случае, если по заданным фильтрам не найдено ни одного события, возвращает пустой список
      operationId: getEvents_2
      parameters:
        - description: список id пользователей, чьи события нужно найти
          in: query
          name: users
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int64
        - description: список состояний в которых находятся искомые события
          in: query
          name: states
          required: false
          schema:
            type: array
            items:
              type: string
        - description: список id категорий в которых будет вестись поиск
          in: query
          name: categories
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int64
        - description: дата и время не раньше которых должно произойти событие
          in: query
          name: rangeStart
          required: false
          schema:
            type: string
        - description: дата и время не позже которых должно произойти событие
          in: query
          name: rangeEnd
          required: false
          schema:
            type: string
        - description: количество событий, которые нужно пропустить для формирования текущего набора
          in: query
          name: from
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - description: количество событий в наборе
          in: query
          name: size
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventFullDto'
          description: События найдены
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: 'Failed to convert value of type java.lang.String to required type int; nested exception is java.lang.NumberFormatException: For input string: ad'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
      summary: Поиск событий
      tags:
        - 'Admin: События'
  /admin/events/{eventId}:
    patch:
      description: |-
        Редактирование данных любого события администратором. Валидация данных не требуется.
        Обратите внимание:
         - дата начала изменяемого события должна быть не ранее чем за час от даты публикации. (Ожидается код ошибки 409)
        - событие можно публиковать, только если оно в состоянии ожидания публикации (Ожидается код ошибки 409)
        - событие можно отклонить, только если оно еще не опубликовано (Ожидается код ошибки 409)
      operationId: updateEvent_1
      parameters:
        - description: id события
          in: path
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventAdminRequest'
        description: Данные для изменения информации о событии
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие отредактировано
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=2 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не найдено или недоступно
        '409':
          content:
            application/json:
              example:
                status: FORBIDDEN
                reason: For the requested operation the conditions are not met.
                message: 'Cannot publish the event because it''s not in the right state: PUBLISHED'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не удовлетворяет правилам редактирования
      summary: Редактирование данных события и его статуса (отклонение/публикация).
      tags:
        - 'Admin: События'
...
  /events:
    get:
      description: |-
        Обратите внимание: 
        - это публичный эндпоинт, соответственно в выдаче должны быть только опубликованные события
        - текстовый поиск (по аннотации и подробному описанию) должен быть без учета регистра букв
        - если в запросе не указан диапазон дат [rangeStart-rangeEnd], то нужно выгружать события, которые произойдут позже текущей даты и времени
        - информация о каждом событии должна включать в себя количество просмотров и количество уже одобренных заявок на участие
        - информацию о том, что по этому эндпоинту был осуществлен и обработан запрос, нужно сохранить в сервисе статистики

        В случае, если по заданным фильтрам не найдено ни одного события, возвращает пустой список
      operationId: getEvents_1
      parameters:
        - description: текст для поиска в содержимом аннотации и подробном описании события
          in: query
          name: text
          required: false
          schema:
            type: string
        - description: список идентификаторов категорий в которых будет вестись поиск
          in: query
          name: categories
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int64
        - description: поиск только платных/бесплатных событий
          in: query
          name: paid
          required: false
          schema:
            type: boolean
        - description: дата и время не раньше которых должно произойти событие
          in: query
          name: rangeStart
          required: false
          schema:
            type: string
        - description: дата и время не позже которых должно произойти событие
          in: query
          name: rangeEnd
          required: false
          schema:
            type: string
        - description: только события у которых не исчерпан лимит запросов на участие
          in: query
          name: onlyAvailable
          required: false
          schema:
            type: boolean
            default: false
        - description: 'Вариант сортировки: по дате события или по количеству просмотров'
          in: query
          name: sort
          required: false
          schema:
            type: string
            enum:
              - EVENT_DATE
              - VIEWS
        - description: количество событий, которые нужно пропустить для формирования текущего набора
          in: query
          name: from
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - description: количество событий в наборе
          in: query
          name: size
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventShortDto'
          description: События найдены
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: Event must be published
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
      summary: Получение событий с возможностью фильтрации
      tags:
        - 'Public: События'
  /events/{id}:
    get:
      description: |-
        Обратите внимание:
        - событие должно быть опубликовано
        - информация о событии должна включать в себя количество просмотров и количество подтвержденных запросов
        - информацию о том, что по этому эндпоинту был осуществлен и обработан запрос, нужно сохранить в сервисе статистики

        В случае, если события с заданным id не найдено, возвращает статус код 404
      operationId: getEvent_1
      parameters:
        - description: id события
          in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие найдено
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: 'Failed to convert value of type java.lang.String to required type int; nested exception is java.lang.NumberFormatException: For input string: ad'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=13 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не найдено или недоступно
      summary: Получение подробной информации об опубликованном событии по его идентификатору
      tags:
        - 'Public: События'
  /users/{userId}/events:
    get:
      description: В случае, если по заданным фильтрам не найдено ни одного события, возвращает пустой список
      operationId: getEvents
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - description: количество элементов, которые нужно пропустить для формирования текущего набора
          in: query
          name: from
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - description: количество элементов в наборе
          in: query
          name: size
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventShortDto'
          description: События найдены
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: 'Failed to convert value of type java.lang.String to required type int; nested exception is java.lang.NumberFormatException: For input string: ad'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
      summary: Получение событий, добавленных текущим пользователем
      tags:
        - 'Private: События'
    post:
      description: 'Обратите внимание: дата и время на которые намечено событие не может быть раньше, чем через два часа от текущего момента'
      operationId: addEvent
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewEventDto'
        description: данные добавляемого события
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие добавлено
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: 'Field: category. Error: must not be blank. Value: null'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '409':
          content:
            application/json:
              example:
                status: FORBIDDEN
                reason: For the requested operation the conditions are not met.
                message: 'Field: eventDate. Error: должно содержать дату, которая еще не наступила. Value: 2020-12-31T15:10:05'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не удовлетворяет правилам создания
      summary: Добавление нового события
      tags:
        - 'Private: События'
  /users/{userId}/events/{eventId}:
    get:
      description: В случае, если события с заданным id не найдено, возвращает статус код 404
      operationId: getEvent
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - description: id события
          in: path
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие найдено
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: 'Failed to convert value of type java.lang.String to required type long; nested exception is java.lang.NumberFormatException: For input string: ad'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=13 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не найдено или недоступно
      summary: Получение полной информации о событии добавленном текущим пользователем
      tags:
        - 'Private: События'
    patch:
      description: |
        Обратите внимание:
        - изменить можно только отмененные события или события в состоянии ожидания модерации (Ожидается код ошибки 409)
        - дата и время на которые намечено событие не может быть раньше, чем через два часа от текущего момента (Ожидается код ошибки 409)
      operationId: updateEvent
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - description: id редактируемого события
          in: path
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventUserRequest'
        description: Новые данные события
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие обновлено
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: Event must not be published
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=283 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не найдено или недоступно
        '409':
          content:
            application/json:
              example:
                status: FORBIDDEN
                reason: For the requested operation the conditions are not met.
                message: Only pending or canceled events can be changed
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не удовлетворяет правилам редактирования
      summary: Изменение события добавленного текущим пользователем
      tags:
        - 'Private: События'
  /users/{userId}/events/{eventId}/requests:
    get:
      description: В случае, если по заданным фильтрам не найдено ни одной заявки, возвращает пустой список
      operationId: getEventParticipants
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - description: id события
          in: path
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParticipationRequestDto'
          description: Найдены запросы на участие
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: 'Failed to convert value of type java.lang.String to required type int; nested exception is java.lang.NumberFormatException: For input string: ad'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
      summary: Получение информации о запросах на участие в событии текущего пользователя
      tags:
        - 'Private: События'
    patch:
      description: |-
        Обратите внимание:
        - если для события лимит заявок равен 0 или отключена пре-модерация заявок, то подтверждение заявок не требуется
        - нельзя подтвердить заявку, если уже достигнут лимит по заявкам на данное событие (Ожидается код ошибки 409)
        - статус можно изменить только у заявок, находящихся в состоянии ожидания (Ожидается код ошибки 409)
        - если при подтверждении данной заявки, лимит заявок для события исчерпан, то все неподтверждённые заявки необходимо отклонить
      operationId: changeRequestStatus
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - description: id события текущего пользователя
          in: path
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventRequestStatusUpdateRequest'
        description: Новый статус для заявок на участие в событии текущего пользователя
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventRequestStatusUpdateResult'
          description: Статус заявок изменён
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: Request must have status PENDING
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=321 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не найдено или недоступно
        '409':
          content:
            application/json:
              example:
                status: CONFLICT
                reason: For the requested operation the conditions are not met.
                message: The participant limit has been reached
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Достигнут лимит одобренных заявок
      summary: Изменение статуса (подтверждена, отменена) заявок на участие в событии текущего пользователя
      tags:
        - 'Private: События'
  /users/{userId}/requests:
    get:
      description: В случае, если по заданным фильтрам не найдено ни одной заявки, возвращает пустой список
      operationId: getUserRequests
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParticipationRequestDto'
          description: Найдены запросы на участие
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: 'Failed to convert value of type java.lang.String to required type long; nested exception is java.lang.NumberFormatException: For input string: ad'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: User with id=11 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Пользователь не найден
      summary: Получение информации о заявках текущего пользователя на участие в чужих событиях
      tags:
        - 'Private: Запросы на участие'
    post:
      description: |-
        Обратите внимание:
        - нельзя добавить повторный запрос  (Ожидается код ошибки 409)
        - инициатор события не может добавить запрос на участие в своём событии (Ожидается код ошибки 409)
        - нельзя участвовать в неопубликованном событии (Ожидается код ошибки 409)
        - если у события достигнут лимит запросов на участие - необходимо вернуть ошибку  (Ожидается код ошибки 409)
        - если для события отключена пре-модерация запросов на участие, то запрос должен автоматически перейти в состояние подтвержденного
      operationId: addParticipationRequest
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - description: id события
          in: query
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipationRequestDto'
          description: Заявка создана
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: 'Failed to convert value of type java.lang.String to required type long; nested exception is java.lang.NumberFormatException: For input string: ad'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=13 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не найдено или недоступно
        '409':
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated.
                message: 'could not execute statement; SQL [n/a]; constraint [uq_request]; nested exception is org.hibernate.exception.ConstraintViolationException: could not execute statement'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Нарушение целостности данных
      summary: Добавление запроса от текущего пользователя на участие в событии
      tags:
        - 'Private: Запросы на участие'
  /users/{userId}/requests/{requestId}/cancel:
    patch:
      operationId: cancelRequest
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - description: id запроса на участие
          in: path
          name: requestId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticipationRequestDto'
          description: Заявка отменена
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Request with id=2727 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос не найден или недоступен
      summary: Отмена своего запроса на участие в событии
      tags:
        - 'Private: Запросы на участие'
...
components:
  schemas:
...
    UpdateEventAdminRequest:
      type: object
      properties:
        annotation:
          maxLength: 2000
          minLength: 20
          type: string
          description: Новая аннотация
          example: Сап прогулки по рекам и каналам – это возможность увидеть Практикбург с другого ракурса
        category:
          type: integer
          description: Новая категория
          format: int64
          example: 3
        description:
          maxLength: 7000
          minLength: 20
          type: string
          description: Новое описание
          example: От английского SUP - Stand Up Paddle — 'стоя на доске с веслом', гавайская разновидность сёрфинга, в котором серфер, стоя на доске, катается на волнах и при этом гребет веслом, а не руками, как в классическом серфинге.
        eventDate:
          type: string
          description: Новые дата и время на которые намечено событие. Дата и время указываются в формате 'yyyy-MM-dd HH:mm:ss'
          example: '2023-10-11 23:10:05'
        location:
          $ref: '#/components/schemas/Location'
        paid:
          type: boolean
          description: Новое значение флага о платности мероприятия
          example: true
        participantLimit:
          type: integer
          description: Новый лимит пользователей
          format: int32
          example: 7
        requestModeration:
          type: boolean
          description: Нужна ли пре-модерация заявок на участие
          example: false
        stateAction:
          type: string
          description: Новое состояние события
          enum:
            - PUBLISH_EVENT
            - REJECT_EVENT
        title:
          maxLength: 120
          minLength: 3
          type: string
          description: Новый заголовок
          example: Сап прогулки по рекам и каналам
      description: Данные для изменения информации о событии. Если поле в запросе не указано (равно null) - значит изменение этих данных не треубется.
...